keep api.js pure, move helpers to their domain code
